==> 01_nlp/01_crawl_hankyoung.ipynb <==
# imports

import pandas as pd
import numpy as np

import requests
from bs4 import BeautifulSoup

# def

def getHankyong(base_url, pages):
    
    articles_list = []
    for page in pages:
        url = base_url + str(page)
        response = requests.get(url)
        
        if response.status_code == 200:
            html = response.text
            soup = BeautifulSoup(html, 'html.parser')
            
            articles = soup.select_one('#container > div.wrap_cont > div.inner_list').select('li')
            for article in articles:
                try:
                    date  = article.select_one('span.time').text
                    title = article.select_one('h3.tit').text
                    url   = article.select_one('h3.tit').select_one('a').get('href')
                    
                    article_row = []
                    article_row.append(page)
                    article_row.append(date)
                    article_row.append(title)
                    article_row.append(url)
                    
                    articles_list.append(article_row)
                except Exception:
                    continue
        else:
            print('error')
    
    df = pd.DataFrame(articles_list, columns=['page', 'date', 'title', 'url'])
    
    return df

# 한국경제 시황

sihwang = getHankyong('https://www.hankyung.com/finance/0101?page=', range(1, 3164))

sihwang

sihwang.to_csv('./hankyoung_sihwang_url.csv', index=False)

sihwang.iloc[0, 3]

import time

def getContent(article_url):
    # time.sleep(10)
    response = requests.get(article_url)
    
    if response.status_code == 200:
        html = response.text
        # soup = BeautifulSoup(html, 'html.parser')
        
        # content = soup.select('#articletxt')
        
        
    else:
        print('response :', response.status_code)
    
    return response

getContent('https://www.hankyung.com/finance/article/2021110949787')

requests.get('https://www.hankyung.com/finance/article/2021111098036')

# 한국경제 해외증시

foreign = getHankyong('https://www.hankyung.com/finance/0105?page=', range(1, 1872))

foreign

foreign.to_csv('./hankyoung_foreign_url.csv', index=False)
