==> 01_nlp/02_crawl_naver.ipynb <==
# imports

import pandas as pd
import numpy as np

import requests
from bs4 import BeautifulSoup

# def

def getNaver(base_url, dates):
    
    articles_list = []
    for date in dates:
        for page in range(1, 21):
            
            url = base_url + str(date) + '&page=' + str(page)
            response = requests.get(url)
            
            if response.status_code == 200:
                html = response.text
                soup = BeautifulSoup(html, 'html.parser')
                
                articles = soup.select_one('#contentarea_left > ul').select('dd.articleSubject')
                for article in articles:
                    try:
                        title       = article.select_one('a').text
                        article_url = 'https://finance.naver.com/' + article.select_one('a').get('href')
                        
                        article_row = []
                        article_row.append(page)
                        article_row.append(date)
                        article_row.append(title)
                        article_row.append(article_url)
                        
                        articles_list.append(article_row)
                        
                    except Exception:
                        continue
            else:
                print('error')
    
    df = pd.DataFrame(articles_list, columns=['page', 'date', 'title', 'url'])
    
    return df

# 네이버 금융 : 시황/전망

sihwang_url = 'https://finance.naver.com/news/news_list.naver?mode=LSS3D&section_id=101&section_id2=258&section_id3=401&date='
dates = pd.date_range('2003-01-01', '2021-11-10').strftime('%Y%m%d').tolist()

sihwang = getNaver(sihwang_url, dates)

sihwang

sihwang.to_csv('./naver_sihwang_url.csv', index=False)

# 네이버 금융 : 해외증시

foreign_url = 'https://finance.naver.com/news/news_list.naver?mode=LSS3D&section_id=101&section_id2=258&section_id3=403&date='
dates = pd.date_range('2003-01-01', '2021-11-10').strftime('%Y%m%d').tolist()

foreign = getNaver(foreign_url, dates)

foreign.to_csv('./nadver_foreign_url.csv', index=False)
