==> 01_nlp/04_article_nlp.ipynb <==
# imports

import pandas as pd
import numpy as np
import re

from nltk.tokenize import sent_tokenize
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
import konlpy

# read

hankyoung_sihwang = pd.read_csv('./input/hankyoung_sihwang_url.csv')
hankyoung_foreign = pd.read_csv('./input/hankyoung_foreign_url.csv')
naver_sihwang     = pd.read_csv('./input/naver_sihwang_url.csv')
naver_foreign     = pd.read_csv('./input/naver_foreign_url.csv')
infonews_sihwang  = pd.read_csv('./input/infonews_국내.csv')
infonews_foreign  = pd.read_csv('./input/infonews_해외.csv')

infonews_sihwang.dropna(inplace=True)

hankyoung_sihwang.shape, hankyoung_foreign.shape, naver_sihwang.shape, naver_foreign.shape, infonews_sihwang.shape, infonews_foreign.shape, 

hankyoung_sihwang['class'] = '한경국내'
hankyoung_foreign['class'] = '한경해외'
naver_sihwang['class']     = '네이버국내'
naver_foreign['class']     = '네이버해외'
infonews_sihwang['class']  = '인포국내'
infonews_foreign['class']  = '인포해외'

hankyoung_sihwang['date'] = hankyoung_sihwang['date'].str[:10]
hankyoung_sihwang['date'] = pd.to_datetime(hankyoung_sihwang['date'])

hankyoung_foreign['date'] = hankyoung_foreign['date'].str[:10]
hankyoung_foreign['date'] = pd.to_datetime(hankyoung_foreign['date'])

naver_sihwang['date'] = pd.to_datetime(naver_sihwang['date'], format='%Y%m%d')
naver_foreign['date'] = pd.to_datetime(naver_foreign['date'], format='%Y%m%d')

infonews_sihwang['date'] = infonews_sihwang['date'].str.split('|').str[-1].str.split(' ').str[-2]
infonews_sihwang['date'] = pd.to_datetime(infonews_sihwang['date'])

infonews_foreign['date'] = infonews_foreign['date'].str.split('|').str[-1].str.split(' ').str[-2]
infonews_foreign['date'] = pd.to_datetime(infonews_foreign['date'])

df = pd.concat(
    [
        hankyoung_sihwang, hankyoung_foreign, 
        naver_sihwang, naver_foreign, 
        infonews_sihwang, infonews_foreign
    ],
)
df.reset_index(drop=True, inplace=True)

df.shape

df.head()

df.tail()

df['date'].max(), df['date'].min()

# df.to_csv('./input/articles_title.csv', index=False)

# Lemmatize

import pandas as pd
import numpy as np
import re

from nltk.tokenize import sent_tokenize
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
import konlpy

df = pd.read_csv('./input/articles_title.csv')

from ckonlpy.tag import Twitter
from nltk.stem import WordNetLemmatizer

tw = Twitter()
lm = WordNetLemmatizer()

titles = df['title'].values.tolist()

headwords = []
for title in titles:
    headwords.append([lm.lemmatize(word) for word in tw.nouns(title)])

headwords[:5]

len(headwords)

# Document-Term Matrix(DTM)

headwords = [' '.join(word) for word in headwords]

headwords[:5]

len(headwords)

df['headwords'] = headwords

# df.to_csv('./input/articles_title_headwords.csv', index=False)
