==> 01_nlp/05_make_dtm.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_rows', 100, 'display.max_columns', 100, 'max_colwidth', None)

df = pd.read_csv('./input/articles_title_headwords.csv')

band00 = '2000-00-00'
band01 = '2006-04-03'
band02 = '2008-11-12'
band03 = '2010-11-10'
band04 = '2012~07~12'
band05 = '2014-04-02'
band06 = '2015-07-26'
band07 = '2016-08-18'
band08 = '2017-07-12'
band09 = '2018-05-07'
band10 = '2018-11-08'
band11 = '2019-05-26'
band12 = '2019-12-18'
band13 = '2020-06-04'
band14 = '2021-02-23'
band15 = '2021-11-12'

band_name = '15'
start_date, end_date = band14, band15
df = df[(df['date'] > start_date) & (df['date'] <= end_date)]

df.shape

from sklearn.feature_extraction.text import CountVectorizer

headwords = df['headwords'].values.tolist()

vector = CountVectorizer()
dtm = vector.fit_transform(headwords).toarray()
dtm_dict = vector.vocabulary_

dtm.shape

idx = pd.DataFrame(dtm_dict, index=['index']).T.sort_values('index').index
dtm_df = pd.DataFrame(dtm, columns=idx)
dtm_count = pd.DataFrame(dtm_df.sum(axis=0), columns=['count']).sort_values('count', ascending=False)

dtm_count = dtm_count[dtm_count['count'] > 50]
dtm_count.reset_index(inplace=True)
dtm_count['band'] = band_name
dtm_count

dtm_count.to_csv('./dtm_count{}.csv'.format(band_name), index=False)
